New branch which supports multiple AWGs.
AWGs are new MATLAB objects with a interface class AWG. The 'driver
class'inherits from this AWG class and implements the hardware specific
methods.

This commit includes the UNTESTED classes for Tektronix and PXDAC AWGs.

The interface is provided by the class VAWG which is a virtual awg and
allows mapping of pulsecontrol channels to harware channels on diffrent
AWGs.